+++ tests/AuthenticatedUser.py
@@ -29,6 +29,7 @@
 # Copyright 2023 Enrico Minack <github@enrico.minack.dev>                      #
 # Copyright 2023 Jirka Borovec <6035284+Borda@users.noreply.github.com>        #
 # Copyright 2024 Chris Wells <ping@cwlls.com>                                  #
+# Copyright 2024 Eduardo Ramírez <edu.rh90@gmail.com>                          #
 # Copyright 2024 Enrico Minack <github@enrico.minack.dev>                      #
 # Copyright 2024 Oskar Jansson <56458534+janssonoskar@users.noreply.github.com>#
 #                                                                              #
@@ -783,3 +784,14 @@ class AuthenticatedUser(Framework.TestCase):
         self.assertEqual(installations[0].target_id, 3344556)
         self.assertEqual(installations[0].target_type, "User")
         self.assertEqual(installations.totalCount, 1)
+
+    def testGetMemberships(self):
+        membership_data = self.user.get_organization_memberships()
+        self.assertListKeyEqual(
+            membership_data,
+            lambda e: e.organization.login,
+            ["aneyem-github", "nko4", "geoservel", "iic2154-uc-cl", "nnodes", "sushiclm"],
+        )
+        self.assertListKeyEqual(
+            membership_data, lambda e: e.role, ["member", "member", "admin", "member", "member", "admin"]
+        )