+++ tests/WorkflowRun.py
@@ -8,6 +8,8 @@
 # Copyright 2023 Jirka Borovec <6035284+Borda@users.noreply.github.com>        #
 # Copyright 2023 Sasha Chung <50770626+nuang-ee@users.noreply.github.com>      #
 # Copyright 2024 Chris Gavin <chris@chrisgavin.me>                             #
+# Copyright 2024 Enrico Minack <github@enrico.minack.dev>                      #
+# Copyright 2024 Geoffrey <geoffrey@moveparallel.com>                          #
 #                                                                              #
 # This file is part of PyGithub.                                               #
 # http://pygithub.readthedocs.io/                                              #
@@ -43,6 +45,7 @@ class WorkflowRun(Framework.TestCase):
             repr(self.workflow_run),
             'WorkflowRun(url="https://api.github.com/repos/PyGithub/PyGithub/actions/runs/3881497935", id=3881497935)',
         )
+        self.assertEqual(self.workflow_run.actor.login, "nuang-ee")
         self.assertEqual(self.workflow_run.id, 3881497935)
         self.assertEqual(self.workflow_run.name, "CI")
         self.assertEqual(self.workflow_run.head_branch, "feat/workflow-run")